Add Column checking.
Allows for columns to be added and removed automaticly at onLoad.

The ablility to check for column attributes like IsPrimaryKey, Type, default and NOT NULL has been removed because SQLite does not grant any easy way to handle these actions.

If it is needed the column has to be renamed, a new column has to be added, the data migrated to the new column, and the old column to be removed.